requirements.txt
# ── Core ML / Vision ─────────────────────────────────────────────────────────
torch
torchvision
transformers
Pillow

# ── Vector Store ──────────────────────────────────────────────────────────────
faiss-cpu

# ── Video Processing ─────────────────────────────────────────────────────────
opencv-python-headless

# ── Numerical ────────────────────────────────────────────────────────────────
numpy

# ── Frontend ─────────────────────────────────────────────────────────────────
streamlit

# ── LLM (optional — works without these) ─────────────────────────────────────
google-generativeai
openai

# ── Utils ────────────────────────────────────────────────────────────────────
python-dotenv
